xen/vpci: Allow function passthrough
Remove the restriction on passing through functions.  vPCI currently
only allows assigning function 0, but we want to allow assigning other
functions.

A single function XX:YY.Z is assigned as funtion 0 for a new device
VV:WW.0.

Passthrough of all function of a multifunction device as a set is not
supported.

xen-pciback passthrough allows passthrough of functions, so it should
behave comparably.

Signed-off-by: Jason Andryuk <jason.andryuk@amd.com>